macOS: Don't send update requests during live resizing
We had a optimization during update request delivery that checked
whether the view or layer needed display, and if so deferred the
update request, under the assumption that the display cycle would
then be the better place to draw the next frame.

Unfortunately on macOS 10.15, the needsDisplay status of the view
may in some cases always be true, which results in us never sending
any update requests.

Removing this optimization is not ideal, as the results of drawing
outside of the display cycle are somewhat glitchy. Instead we opt
to be even more conservative than the original intent, and defer
all update requests during live resizing.

This has the effect of stopping all animations while a window is
resized. Clients who prefer the animations to continue, with the
possible glitches that may come with it, can set the unsupported
QT_MAC_DELIVER_UPDATE_REQUESTS_DURING_RESIZE environment variable.
This variable will go away once the situation has been diagnosed
further and deemed unnecessary.

[ChangeLog][macOS] Update requests will not be delivered during
live resizing, to prevent severe visual artifacts of doing so.

Change-Id: Iddc436cbd8c6411a33f2999df9c1e982fd1898bf
Fixes: QTBUG-77958